+++ fractal_server/app/models/v2/workflowtask.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from pydantic import ConfigDict
 from sqlalchemy import Column
-from sqlalchemy.types import JSON
+from sqlalchemy.dialects.postgresql import JSONB
 from sqlmodel import Field
 from sqlmodel import Relationship
 from sqlmodel import SQLModel
@@ -18,20 +18,20 @@ class WorkflowTaskV2(SQLModel, table=True):
     workflow_id: int = Field(foreign_key="workflowv2.id", ondelete="CASCADE")
     order: int | None = None
     meta_parallel: dict[str, Any] | None = Field(
-        sa_column=Column(JSON), default=None
+        sa_column=Column(JSONB), default=None
     )
     meta_non_parallel: dict[str, Any] | None = Field(
-        sa_column=Column(JSON), default=None
+        sa_column=Column(JSONB), default=None
     )
     args_parallel: dict[str, Any] | None = Field(
-        sa_column=Column(JSON), default=None
+        sa_column=Column(JSONB), default=None
     )
     args_non_parallel: dict[str, Any] | None = Field(
-        sa_column=Column(JSON), default=None
+        sa_column=Column(JSONB), default=None
     )
 
     type_filters: dict[str, bool] = Field(
-        sa_column=Column(JSON, nullable=False, server_default="{}")
+        sa_column=Column(JSONB, nullable=False, server_default="{}")
     )
 
     # Task